Revert to `unconsM`.
