+++ src/Trisagion/Parsers/Streamable.hs
@@ -9,6 +9,7 @@ module Trisagion.Parsers.Streamable (
     ValidationError (..),
 
     -- * Parsers @'Streamable' m a s => 'ParserT' m s e a@.
+    one,
     skipOne,
     peek,
     satisfy,
@@ -22,10 +23,28 @@ import Data.Functor (($>))
 import Data.Functor.Identity (Identity (..))
 import Data.Void (Void)
 
+-- Libraries.
+import Control.Monad.State (MonadState (..))
+import Control.Monad.Trans (MonadTrans (..))
+
+-- non-Hackage libraries.
+import Mono.Typeclasses.MonoFunctor (MonoFunctor (..))
+
 -- Package.
 import Trisagion.Types.Either ((:+:))
-import Trisagion.Typeclasses.Streamable (Streamable (..), InputError (..))
-import Trisagion.ParserT (ParserT, lookAhead, try, validate)
+import Trisagion.Typeclasses.Streamable (Streamable (..))
+import Trisagion.ParserT (ParserT, lookAhead, try, validate, throw)
+
+
+{- | The t'InputError' error type. -}
+newtype InputError = InputError Word
+    deriving stock (Eq, Show)
+
+{- | Monofunctoriality of t'InputError'. -}
+instance MonoFunctor Word InputError where
+    {-# INLINE monomap #-}
+    monomap :: (Word -> Word) -> InputError -> InputError
+    monomap f (InputError n) = InputError (f n)
 
 
 {- | The t'ValidationError' error tag type thrown on failed validations. -}
@@ -33,6 +52,14 @@ newtype ValidationError e = ValidationError e
     deriving stock (Eq, Show, Functor, Foldable, Traversable)
     deriving (Applicative, Monad) via Identity
 
+
+{- | Parse one element from the input stream. -}
+one :: Streamable m a s => ParserT s InputError m a
+one = do
+    r <- get >>= lift . unconsM
+    case r of
+        Nothing      -> throw (InputError 1)
+        Just (x, xs) -> put xs $> x
 
 {- | Skip one element from the input stream. -}
 {-# INLINE skipOne #-}
@@ -49,7 +76,7 @@ peek = either (const Nothing) Just <$> lookAhead one
 satisfy
     :: forall m a s . Streamable m a s
     => (a -> Bool)                      -- ^ Predicate on @a@.
-    -> ParserT s ((ValidationError a) :+: InputError) m a
+    -> ParserT s (ValidationError a :+: InputError) m a
 satisfy p = validate v one
     where
         v :: a -> ValidationError a :+: a
@@ -59,14 +86,14 @@ satisfy p = validate v one
 {-# INLINE single #-}
 single
     :: (Streamable m a s, Eq a)
-    =>a                                -- ^ Matching @x :: a@.
-    -> ParserT s ((ValidationError a) :+: InputError) m a
+    => a                                -- ^ Matching @x :: a@.
+    -> ParserT s (ValidationError a :+: InputError) m a
 single x = satisfy (== x)
 
 {- | Parse one element from the input stream @'Streamable' m a s@ matching an @x :: a@. -}
 {-# INLINE oneOf #-}
 oneOf
     :: (Streamable m a s, Eq a, Foldable t)
-    =>t a                              -- ^ Foldable of @x :: a@ against which to test inclusion.
-    -> ParserT s ((ValidationError a) :+: InputError) m a
+    => t a                              -- ^ Foldable of @x :: a@ against which to test inclusion.
+    -> ParserT s (ValidationError a :+: InputError) m a
 oneOf xs = satisfy (`elem` xs)
